feat(simulated annealing): Judment and Mutation Updating code

+++ src/main/java/com/SCIA/SimulatedAnneling/InitialHeat.java
@@ -7,11 +7,11 @@ import com.SCIA.Schedule.Event.Event;
 
 public class InitialHeat {
 
-    public static float initialHeat(List<Event> eventList)  {
+    public static double initialHeat(List<Event> eventList)  {
         int conflicts = Judgement.getConflicts(eventList);
 
         //calculate heat so maybe the conflicts value times 100000
-        float return_value = conflicts*100000;
+        double return_value = conflicts*100000;
 
         return return_value;
     }
