+++ src/main/java/com/SCIA/SimulatedAnneling/Probability.java
@@ -1,7 +1,7 @@
 package com.SCIA.SimulatedAnneling;
 
 public class Probability {
-    public static double getProbability(float sigma, int conflicts)    {
+    public static double getProbability(double sigma, int conflicts)    {
         double probability = Math.exp(-conflicts/sigma);
         return probability;
     }
